requirements.txt
numpy>=1.15.2
scipy>=1.1.0
lxml>=4.2.5
commonroad-io>=2018.1

# only for graphical display
PyQt5==5.11.3
matplotlib>=3.0.0
